Minimize number of syscalls by ~3x while accepting connections
Go back to the model where epoll_ctl(ADD, client_fd) is performed,
nudging the thread after a batch of clients have been accepted.

Before:

	for every client {
		write(pipe, client_fd)
		read(pipe)
		epoll_ctl(epoll_fd, ADD, client_fd)
	}

After:

	for every client {
		epoll_ctl(epoll_fd, ADD, client_fd)
	}
	write(pipe, 1)
	read(pipe)

Only threads that had connections scheduled to them will be "nudged";
this is saved in a 64-bit integer, so this limits the number of threads
to 64.  Not a big deal.

These gains can only be reaped in full when a herd of clients connect;
otherwise, not much will actually change.

The thread I/O loop had to be refactored as well, which is itself a
good bonus.  This should make it easier to implement timers now
(although the SPSC queue might need to be replaced by a MPSC queue;
we shall see).